% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityFunctions.R
\name{activityVisualise}
\alias{activityVisualise}
\title{Visualise Activity data}
\usage{
activityVisualise(
  DataFrame,
  DataType,
  TimeBreak = "no",
  PageNumber = NA_real_,
  StudyID
)
}
\arguments{
\item{DataFrame}{A dataframe containing Fitbit activity data.}

\item{DataType}{Character object determining which data to visualise. 2 options available: "stepcount"
or "heartrate".}

\item{TimeBreak}{Character object which defines whether plot outputs should be split
by time period. Default is "No" in which case there will be a single plot produce
including data for the whole period. Other options are "week" and "day" in which case multiple
plots will be produced according to TimeBreak.}

\item{PageNumber}{Vector indicating which page (i.e. week/day) number selected for
visualisation as plot is faceted according to TimeBreak.}

\item{StudyID}{ID of participant for whom step count or heart rate data will be plotted.}
}
\value{
Bar chart displaying step count or heart rate over time
}
\description{
Creates a bar chart of the number of steps or heart rate
minute by minute over the study period
}
\examples{

\dontrun{
hypometrics::activityVisualise(DataFrame,
                               DataType = "stepcount",
                               TimeBreak = "no",
                               PageNumber = 1,
                               StudyID = "01001")
}

}
