% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cgm}
\alias{cgm}
\title{Dataset containing glucose data of 3 participants
who used a CGM device}
\format{
A data frame with one row per CGM timestamp
\describe{
    \item{id}{Unique study ID}
    \item{cgm_timestamp}{Timestamp indicated on CGM device}
    \item{glucose}{Glucose concentration (mmol/L)}
}
}
\usage{
cgm
}
\description{
Dataset containing glucose data of 3 participants
who used a CGM device
}
\keyword{datasets}
