% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkingFunctions.R
\name{cgmactivityLink}
\alias{cgmactivityLink}
\title{Links Continuous Glucose Monitoring (CGM) Data with Activity Data}
\usage{
cgmactivityLink(CgmDataFrame, ActivityDataFrame, DataType)
}
\arguments{
\item{CgmDataFrame}{A dataframe containing CGM data. Must have columns id, cgm_timestamp, glucose.}

\item{ActivityDataFrame}{A dataframe containing Fitbit step count or heart rate data.}

\item{DataType}{Character object determining which data to process. 2 options available: "stepcount"
or "heartrate".}
}
\value{
A dataset where for each CGM timestamp there is a corresponding step count or mean heart rate.
}
\description{
Creates a dataset containing CGM data matched with activity data:
for each CGM timestamp, mean step count and/or heart rate is available
}
\examples{

\dontrun{
hypometrics::cgmactivityLink(CgmDataFrame,
                             ActivityDataFrame,
                             DataType)
}

}
