% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgmInterpolate.R
\name{cgmInterpolate}
\alias{cgmInterpolate}
\title{cgmInterpolate}
\usage{
cgmInterpolate(DataFrame)
}
\arguments{
\item{DataFrame}{A dataframe of CGM data which will be interpolated to one minute epochs.
Must contain columns: id, time, gl.}
}
\value{
A dataframe of CGM data interpolated to 1 minute epochs.
}
\description{
Interpolate CGM values down to one minute epochs.
}
\examples{
\dontrun{
hypometrics::cgmInterpolate(DataFrame)
}
}
