% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkingFunctions.R
\name{cgmprhLink}
\alias{cgmprhLink}
\title{Links Continuous Glucose Monitoring (CGM) Time Series Data with Person-Reported Hypoglycaemia (PRH) Data}
\usage{
cgmprhLink(CgmDataFrame, PrhDataFrame)
}
\arguments{
\item{CgmDataFrame}{A dataframe containing CGM data. Must have columns id, cgm_timestamp, glucose.}

\item{PrhDataFrame}{A dataframe containing all PRH episodes reported with a time. Output of the \link[hypometrics]{umotifLink}
function.}
}
\value{
A dataset containing CGM data with timestamp of PRH episode aligned with closest CGM timestamp.
}
\description{
Creates a dataset containing longitudinal CGM time series data with PRH episodes
on the same row as the closest (in time) CGM timestamps. To run this function, the PRH dataset must be first built
using the prhLink function.
}
\examples{

\dontrun{
hypometrics::cgmprhLink(CgmDataFrame = cgm,
                        PrhDataFrame = prh_map)
}

}
