% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkingFunctions.R
\name{cgmsleepLink}
\alias{cgmsleepLink}
\title{Links Continuous Glucose Monitoring (CGM) Data with Sleep Data}
\usage{
cgmsleepLink(CgmDataFrame, SleepDataFrame)
}
\arguments{
\item{CgmDataFrame}{A dataframe containing CGM data. Must have columns id, cgm_timestamp, glucose.}

\item{SleepDataFrame}{A dataframe containing Fitbit sleep data.}
}
\value{
A dataset with every cgm timestamp marked as
either asleep (timestamp fell within sleep start and and time),
awake (timestamp was outside sleep interval) or
NA (no sleep information available at the time)
}
\description{
Creates a dataset containing CGM data matched with sleep data:
for each CGM timestamp, sleep status is available
}
\examples{

\dontrun{
hypometrics::cgmsleepLink(CgmDataFrame = cgm,
                          SleepDataFrame)
}

}
