% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgmFunctions.R
\name{cgmVisualise}
\alias{cgmVisualise}
\title{Visualise Glucose Data}
\usage{
cgmVisualise(DataFrame, StudyID, TimeBreak = "no", PageNumber = NA_real_)
}
\arguments{
\item{DataFrame}{A dataframe of CGM data which will be filled and interpolated
where specified. Must contain columns: id, cgm_timestamp and glucose.}

\item{StudyID}{ID of participant for whom CGM data will be plotted.}

\item{TimeBreak}{Character object which defines whether plot outputs should be split
by time period. Default is "No" in which case there will be a single plot produce
including data for the whole period. Other options are "week" and "day" in which case multiple
plots will be produced according to TimeBreak.}

\item{PageNumber}{Vector indicating which page (i.e. week/day) number selected for
visualisation as plot is faceted according to TimeBreak.}
}
\value{
A graphical representation showing glucose trace over time with shaded area
representing night time.
}
\description{
Plots Glucose Data Over Time
}
\details{
This functions plots CGM data over time with grey shaded area corresponding to the time period
between 00:00 and 06:00 as typically used to describe nocturnal hypoglycaemia. The function
offers the options to look at the  glucose data over the entire study period, for a specific week or
specific date.
}
\examples{
\dontrun{
hypometrics::cgmVisualise(DataFrame,
                          StudyID = "001",
                          TimeBreak = "day",
                          PageNumber = 7)
}

}
