% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMissingCGM.R
\name{checkMissingCGM}
\alias{checkMissingCGM}
\title{checkMissingCGM}
\usage{
checkMissingCGM(DataFrame)
}
\arguments{
\item{DataFrame}{A dataframe containing CGM data in which missingness will be assessed.
Must have columns id, time, gl.}
}
\value{
A list containing: the CGM data with an additional Date and total_cgm_hours_per_day column,
a visualisation of the distribution of NAs within the data, a visualisation of the percent of NA's
in the data, and a visualisation of the pattern of NAs in the data.
}
\description{
Calculates the completeness of CGM and provides visualisations to examine the amount of missing data.
}
\examples{
\dontrun{
hypometrics::checkMissingCGM(DataFrame)
}

}
