% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umotifFunctions.R
\name{prhSummarise}
\alias{prhSummarise}
\title{Key Person-Reported Hypoglycaemia Metrics}
\usage{
prhSummarise(DataFrame, AddSleepSummary = "no")
}
\arguments{
\item{DataFrame}{A dataframe object, output of the \code{\link[hypometrics]{prhLink}()} function.
It includes for each participant, one hypoglycaemic episode per row with characteristics in each column.}

\item{AddSleepSummary}{A character object specifying whether a summary of PRH should be added according to
sleep status. Default is "no". Other option is "yes".}
}
\value{
Data frame with one line per participant with the count of different types of PRHs.
This includes the number of symptomatic, prevented episodes, night status, and sleep status where specified.
}
\description{
Summarises person-reported hypoglycaemia (PRH) data
}
\details{
This function goes hand in hand with the \code{\link[hypometrics]{prhLink}()} function. Once
the details of each individual hypoglycaemia episode has been produced, this function can be used to summarise
them. There is an option to obtain a summary according to sleep status if the AddSleepSummary is set as "yes".
}
\examples{
\dontrun{
hypometrics::prhSummarise(DataFrame,
                          AddSleepSummary = "yes")
}

}
