% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_checkin}
\alias{raw_checkin}
\title{Dataset containing raw morning questionnaire data of 2 participants
who used the Hypo-METRICS app on the uMotif platform}
\format{
A data frame with one row per questionnaire completed
}
\usage{
raw_checkin
}
\description{
Dataset containing raw morning questionnaire data of 2 participants
who used the Hypo-METRICS app on the uMotif platform
}
\details{
Dataset containing timing of hypoglycaemic events reported
in the daily app questionnaires

\describe{
    \item{userid}{Unique uMotif ID}
    \item{stage}{Study day when questionnaire was completed}
    \item{localTimestamp}{Timestamp of the submission of the questionnaire}
    \item{AM_forfirsthypoAtwhattimedidthishappen}{Timestamp of the first hypo since last questionnaire}
    \item{AM_forsecondhypoAtwhattimedidthishappen}{Timestamp of the second hypo since last questionnaire}
    \item{AM_forthirdhypoAtwhattimedidthishappen}{Timestamp of the third hypo since last questionnaire}
    \item{AM_forfourthhypoAtwhattimedidthishappen}{Timestamp of the fourth hypo since last questionnaire}
    \item{AM_forfifthhypoAtwhattimedidthishappen}{Timestamp of the fifth hypo since last questionnaire}
    \item{AM_forfirsthypoIdontknowatwhattimedidthishappen}{Whether participant does not know when first hypo occurred}
    \item{AM_forsecondhypoIdontknowatwhattimedidthishappen}{Whether participant does not know when second hypo occurred}
    \item{AM_forthirdhypoIdontknowatwhattimedidthishappen}{Whether participant does not know when third hypo occurred}
    \item{AM_forfourthhypoIdontknowatwhattimedidthishappen}{Whether participant does not know when fourth hypo occurred}
    \item{AM_forfirsthypoHowdidyoudetectyourhypoorahypothatwasabouttohappen}{How did participant detect first hypo}
    \item{AM_forsecondhypoHowdidyoudetectyourhypoorahypothatwasabouttohappen}{How did participant detect first hypo}
    \item{AM_forthirdhypoHowdidyoudetectyourhypoorahypothatwasabouttohappen}{How did participant detect first hypo}
    \item{AM_forfourthhypoHowdidyoudetectyourhypoorahypothatwasabouttohappen}{How did participant detect first hypo}
    \item{AM_forfifthhypoHowdidyoudetectyourhypoorahypothatwasabouttohappen}{How did participant detect first hypo}
    \item{AM_forfirsthypoHowdidyoudetectyourhypoorahypothatwasabouttohappenotherpleasespecify}{Other hypo detection - please specify}
    \item{AM_forfirsthypoHowdidyoudetectyourhypoorahypothatwasabouttohappenotherpleasespecify}{Other hypo detection - please specify}
    \item{AM_forfirsthypoHowdidyoudetectyourhypoorahypothatwasabouttohappenotherpleasespecify}{Other hypo detection - please specify}
    \item{AM_forfirsthypoWhathappened}{What happened for first hypo}
    \item{AM_forsecondhypoWhathappened}{What happened for second hypo}
    \item{AM_forthirdhypoWhathappened}{What happened for third hypo}
    \item{AM_forfourthhypoWhathappened}{What happened for fourth hypo}
    \item{AM_forfifthypoWhathappened}{What happened for fifth hypo}
}
}
\keyword{datasets}
