% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_motif_segment}
\alias{raw_motif_segment}
\title{Dataset containing raw umotif data of 2 participants
who used the Hypo-METRICS app on the uMotif platform}
\format{
A data frame with one row per motif segment per participant
}
\usage{
raw_motif_segment
}
\description{
Dataset containing raw umotif data of 2 participants
who used the Hypo-METRICS app on the uMotif platform
}
\details{
Dataset containing details on person-reported hypoglycaemic events

\describe{
    \item{UserId}{Unique uMotif ID}
    \item{uMotifTime}{Timestamp of the submission of the hypo episode on the app}
    \item{SegmentID}{Motif segment ID/code e.g. 1 for "Hypo time", 2 for "Glucose level"}
    \item{Value}{Code for response to segment e.g. 1 for "Now" or "<2mmol/L",
    2 for "15 min ago" or "2-2.9mmol/L"}
}
}
\keyword{datasets}
