% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_sleep}
\alias{raw_sleep}
\title{Dataset containing raw sleep data of 2 participants
who used a Fitbit Charge 4}
\format{
A data frame with one row per sleep event
}
\usage{
raw_sleep
}
\description{
Dataset containing raw sleep data of 2 participants
who used a Fitbit Charge 4
}
\details{
Dataset containing Fitbit Charge 4 sleep information

\describe{
    \item{id}{Unique study ID}
    \item{logId}{Unique sleep event ID}
    \item{dateOfSleep}{Date of the sleep event}
    \item{startTime}{Start time of sleep (YYYY-MM-DD HH:MM:SS)}
    \item{endTime}{End time of sleep (YYYY-MM-DD HH:MM:SS)}
}
}
\keyword{datasets}
