% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_step}
\alias{raw_step}
\title{Dataset containing raw step count data of 2 participants
who used a Fitbit Charge 4}
\format{
A data frame with one row per minute
}
\usage{
raw_step
}
\description{
Dataset containing raw step count data of 2 participants
who used a Fitbit Charge 4
}
\details{
Dataset containing Fitbit Charge 4 step count

\describe{
    \item{id}{Unique study ID}
    \item{step_timestamp}{Timestamp}
    \item{count}{Step count}
}
}
\keyword{datasets}
