% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_wpai}
\alias{raw_wpai}
\title{Dataset containing raw WPAI questionnaire data of 2 participants
who used the Hypo-METRICS app on the uMotif platform}
\format{
A data frame with one row per questionnaire submitted
}
\usage{
raw_wpai
}
\description{
Dataset containing raw WPAI questionnaire data of 2 participants
who used the Hypo-METRICS app on the uMotif platform
}
\details{
Dataset containing responses to weekly WPAI questionnaire

\describe{
    \item{userid}{Unique uMotif ID}
    \item{stage}{Study day when questionnaire was completed}
    \item{localTimestamp}{Timestamp of the submission of the questionnaire}
    \item{areyoucurrentlyemployedworkingforpay}{No/Yes}
    \item{duringthepastsevendayshowmanyhoursdidyouactuallywork}{Number field}
    \item{duringthepastsevendayshowmuchdidhypoglycaemiaaffectyourabilitytodoyourregulardailyactivitiesotherthanworkatajob}{Number field}
    \item{duringthepastsevendayshowmuchdidhypoglycaemiaaffectyourproductivitywhileyouwereworking}{Number field}
    \item{duringthepastsevendayshowmanyhoursdidyoumissfromworkbecauseofanyotherreasonsuchasvacationholidaystimeofftoparticipateinthisstudy}{Scale 0-10}
    \item{duringthepastsevendayshowmanyhoursdidyoumissfromworkbecauseofproblemsassociatedwithyourhypoglycaemia}{Scale 0-10}
}
}
\keyword{datasets}
