% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhDetection.R
\name{sdhDetection}
\alias{sdhDetection}
\title{sdhDetection}
\usage{
sdhDetection(DataFrame, DetectionLimit, DetectionDuration)
}
\arguments{
\item{DataFrame}{A dataframe of CGM in which hypoglycaemic episodes will be detected.
Must have columns id, cgm_timestamp, glucose.}

\item{DetectionLimit}{The glucose value used to detect hypoglycaemia}

\item{DetectionDuration}{The amount of time an individual needs to be at or below the DetectionLimit for a hypolgycaemic episode to be detected}
}
\value{
A dataframe with one hypoglycaemic episode per row with characteristics in each column
}
\description{
Function which summarises raw CGM data and produces an output which
includes key characteristics (e.g nadir, duration) of each hypoglycaemic episode
}
