% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkingFunctions.R
\name{sdhprhLink}
\alias{sdhprhLink}
\title{Links Sensor-Detected Hypoglycaemia (SDH) with Person-Reported Hypoglycaemia (PRH) episodes}
\usage{
sdhprhLink(SdhDataFrame, PrhDataFrame)
}
\arguments{
\item{SdhDataFrame}{A dataframe containing all SDH episodes. Output of the \link[hypometrics]{sdhDetection}
function.}

\item{PrhDataFrame}{A dataframe containing all PRH episodes reported with a time. Output of the \link[hypometrics]{umotifLink}
function.}
}
\value{
A dataset containing both SDHs and PRHs in chronological order
for each participant. Where PRHs fall within -/+ 1h of SDH interval,
those have been matched (i.e considered to be the same event and shown on a unique row).
}
\description{
Creates a dataset containing SDH episodes (from CGM data)
linked with PRH episodes (from uMotif data). There is one row per hypoglycaemic episode.
When PRH episodes fall within  +/- 1h the SDH interval, they are reported on the same row.
To run this function, SDH and PRH datasets must be first built using sdhDetection and umotifprhLink functions.
}
\examples{

\dontrun{
hypometrics::sdhprhLink(SdhDataFrame = sdh,
                        PrhDataFrame = prh)
}

}
