% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgmMetrics.R
\name{sdhSummarise}
\alias{sdhSummarise}
\title{Key Sensor-Detected Hypoglycaemia Metrics}
\usage{
sdhSummarise(DataFrame, DetectionLimit = "3.9", LongDuration = 120)
}
\arguments{
\item{DataFrame}{A dataframe object, output of the \code{\link[hypometrics]{sdhDetection}()} function.
It includes for each participant, one hypoglycaemic episode per row with characteristics in each column.}

\item{DetectionLimit}{Object of type numeric or integer corresponding to
the glucose value used to detect hypoglycaemia. Default is 3.9 mmol/L.}

\item{LongDuration}{Numeric object indicating the minimum duration used to define a long
episode of sensor detected hypoglycaemia. Default is 120 minutes.}
}
\value{
Data frame with one line per participant with key SDH metrics. This includes the number of episodes
during the day/night, mean duration of episodes, the number long SDH
and the number of days during which those occur.
}
\description{
Summarise sensor-detected hypoglycaemia (SDH) data for the chosen glucose threshold
}
\details{
This function goes hand in hand with the \code{\link[hypometrics]{sdhDetection}()} function. Once
the details of each individual hypoglycaemia episode has been produced, this function can be used to summarise
them.
}
\examples{
\dontrun{
hypometrics::sdhSummarise(DataFrame,
                          DetectionLimit = "3.0",
                          LongDuration = 120)
}

}
