% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepFunctions.R
\name{sleepCategorise}
\alias{sleepCategorise}
\title{Categorise Sleep Information from Fitbit Raw Data}
\usage{
sleepCategorise(DataFrame)
}
\arguments{
\item{DataFrame}{A dataframe containing Fitbit sleep data.}
}
\value{
Three dataframes: one including information on main sleeping periods,
another containing more detailed data regarding sleeping stages and the last one including
information on short wake periods occurring during the night.
}
\description{
Function creates 3 separate datasets according to Fitbit sleep data:
main sleeping periods, short wake periods and detailed sleep stages.
}
\examples{

\dontrun{
hypometrics::sleepCategorise(DataFrame)
}

}
