% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepFunctions.R
\name{sleepRead}
\alias{sleepRead}
\title{Unzip, Read and Combine Raw Sleep Files Downloaded
From User's Fitbit account}
\usage{
sleepRead(Unzip = FALSE, FolderPath, FileType, StudyID)
}
\arguments{
\item{Unzip}{Logical string (TRUE/FALSE) which determines whether fitbit folder needs to be
unzipped or not}

\item{FolderPath}{Character object indicating path to folder where Fitbit data is stored}

\item{FileType}{Character object indicating what type of file is to be read. Can be either
"json" or "csv".}

\item{StudyID}{ID of participant for whom activity data will be read.}
}
\value{
A dataset containing original Fitbit sleep
data with an additional column with participant's ID.
}
\description{
This function enables the unzipping, reading and
combining of raw sleep files following download from user's
Fitbit account.
}
\examples{

\dontrun{
hypometrics::sleepRead(Unzip = TRUE,
                       FolderPath = "~/Documents",
                       FileType = "json",
                       StudyID = "001")
}

}
