% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepFunctions.R
\name{sleepSummarise}
\alias{sleepSummarise}
\title{Summarises Sleep Information from Fitbit Raw Data}
\usage{
sleepSummarise(DataFrame)
}
\arguments{
\item{DataFrame}{A dataframe containing Fitbit sleep data.}
}
\value{
A summarised dataset with average time in bed, asleep awake as well as
number of nights with or without missing sleep data for each participant
}
\description{
Creates a dataset containing summarised sleep data such as
mean time in bed, sleep duration and time awake
}
\examples{

\dontrun{
hypometrics::sleepSummarise(DataFrame)
}

}
