% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepFunctions.R
\name{sleepVisualise}
\alias{sleepVisualise}
\title{Visualise Sleeping Patterns}
\usage{
sleepVisualise(DataFrame, VisualiseAll = TRUE, StudyID = "")
}
\arguments{
\item{DataFrame}{A dataframe containing Fitbit sleep data.}

\item{VisualiseAll}{Logical string (TRUE/FALSE) which determines whether histograms will be plotted for all participants
or a selected participant. Default is TRUE.}

\item{StudyID}{ID of participant for whom sleep patterns will be plotted. This is only
relevant if VisualiseAll = FALSE, and will produce individualised graphs.}
}
\value{
Two histogram plots displaying all times recorded by the Fitbit
when participant(s) went to bed and got up.
}
\description{
Creates histograms showin the distribution of times when
participant(s) went to bed and got up
}
\examples{

\dontrun{
hypometrics::sleepVisualise(DataFrame,
                            VisualiseAll = TRUE,
                            StudyID = "")
}

}
