% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umotifFunctions.R
\name{umotifClean}
\alias{umotifClean}
\title{Cleans Raw uMotif Files Downloaded From the Data Download Portal}
\usage{
umotifClean(DataFrame, FileType)
}
\arguments{
\item{DataFrame}{A dataframe containing uMotif data.}

\item{FileType}{Character object indicating what type of file is to be cleaned. Can be either
"motif", "checkin", "promis", "wpai" or "eq5d5l".}
}
\value{
A dataset containing clean uMotif data (e.g. daily questionnaires, wpai)
}
\description{
This function enables the cleaning of raw uMotif data according to the type of file
}
\examples{

\dontrun{
hypometrics::umotifClean(DataFrame = raw_checkin,
                        FileType = "checkin")
}

}
