% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umotifFunctions.R
\name{umotifRead}
\alias{umotifRead}
\title{Unzips and Reads Raw uMotif Files Downloaded From the Data Download Portal}
\usage{
umotifRead(Unzip = FALSE, FolderPath, FilePattern)
}
\arguments{
\item{Unzip}{Logical string (TRUE/FALSE) which determines whether uMotif folder needs to be
unzipped or not}

\item{FolderPath}{Character object indicating path to folder where uMotif data is stored}

\item{FilePattern}{Character object indicating the pattern in the file name that will be used to
extract the uMotif files of interest. For example, "evening-checkin" for evening questionnaires,
"wpai" for work productivity questionnaire, "motif_segmentvalue" for symptoms files.}
}
\value{
A dataset containing original uMotif data (e.g. daily questionnaires, wpai)
}
\description{
This function enables the unzipping and reading of raw uMotif csv files
following download from data download portal.
}
\examples{

\dontrun{
hypometrics::umotifRead(Unzip = TRUE,
                        FolderPath = "~/Documents",
                        FilePattern = "morning-checkin"
                        )
}

}
