% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_eq5d5l}
\alias{raw_eq5d5l}
\title{Dataset containing raw EQ5D5L questionnaire data of 2 participants
who used the Hypo-METRICS app on the uMotif platform}
\format{
A data frame with one row per questionnaire submitted
}
\usage{
raw_eq5d5l
}
\description{
Dataset containing raw EQ5D5L questionnaire data of 2 participants
who used the Hypo-METRICS app on the uMotif platform
}
\details{
Dataset containing responses to daily EQ5D5L questionnaire  on scale from 1 to 5

\describe{
    \item{userid}{Unique uMotif ID}
    \item{stage}{Study day when questionnaire was completed}
    \item{localTimestamp}{Timestamp of the submission of the questionnaire}
    \item{SC}{Self Care}
    \item{AD}{Anxiety/depression}
    \item{UA}{Usual activities}
    \item{MB}{Mobility}
    \item{PD}{Pain/discomfort}
}
}
\keyword{datasets}
