% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_hr}
\alias{raw_hr}
\title{Dataset containing raw heart rate data of 2 participants
who used a Fitbit Charge 4}
\format{
A data frame with one row per minute
}
\usage{
raw_hr
}
\description{
Dataset containing raw heart rate data of 2 participants
who used a Fitbit Charge 4
}
\details{
Dataset containing Fitbit Charge 4 heart rate

\describe{
    \item{id}{Unique study ID}
    \item{hr_timestamp}{Timestamp}
    \item{count}{Step count}
}
}
\keyword{datasets}
