% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_promis}
\alias{raw_promis}
\title{Dataset containing raw promis questionnaire data of 2 participants
who used the Hypo-METRICS app on the uMotif platform}
\format{
A data frame with one row per questionnaire submitted
}
\usage{
raw_promis
}
\description{
Dataset containing raw promis questionnaire data of 2 participants
who used the Hypo-METRICS app on the uMotif platform
}
\details{
Dataset containing responses to promis questionnaire on Likert scale
from 1 to 5.

\describe{
    \item{userid}{Unique uMotif ID}
    \item{stage}{Study day when questionnaire was completed}
    \item{localTimestamp}{Timestamp of the submission of the questionnaire}
    \item{Sleep115}{I was satisfied with my sleep}
    \item{Sleep116}{My sleep was refreshing}
    \item{Sleep90}{I had trouble sleeping}
    \item{Sleep110}{I got enough sleep}
    \item{Sleep44}{I had difficulty falling asleep}
    \item{Sleep87}{I had trouble staying asleep}
    \item{Sleep108}{My sleep was restless}
    \item{Sleep109}{My sleep quality was...}
}
}
\keyword{datasets}
